Fix access and mutation of special private attributes.
Ensures correct function with slotted dataclasses on Python 3.11+. Did
not carefully read the '__slots__' specification in the Python data
model documentation. Neither did the implementers of dataclasses in the
Python standard library; this meant that their implementation surfaced
parent slots on a derived class. A patch to Python 3.11.0 alpha 7 fixed
dataclasses so that they did not surface parent slots anymore
(https://github.com/python/cpython issue 90540). However, we only ran
doctests against Python 3.10 until recently and did not notice the
failures in Python 3.11+. Changing doctests to run as part of the test
suite instead of during Sphinx generation exposed this oversight since
the test suite is run across a matrix of platforms and versions.